backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
asyncpg

python-jose[cryptography]
passlib[bcrypt]

pydantic-settings
python-dotenv
pydantic[email]

argon2-cffi